requirements.txt
jiwer
rouge-score
requests
fastapi
uvicorn[standard]
python-dotenv